requirements.txt
gradio>=5.33.0
gradio_client>=1.10.2
ollama>=0.5.1
pypdf>=4.2.0
pdf2image>=1.17.0
Pillow>=10.3.0
llama-index>=0.12.23
llama-index-llms-ollama>=0.6.2
llama-index-embeddings-ollama>=0.6.0
ipykernel
nvitop
torch
torchvision
